requirements.txt
appnope==0.1.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.1
decorator==4.3.0
Django==2.1.2
django-cors-middleware==1.3.1
django-extensions==2.1.3
django-nose==1.4.6
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
docopt==0.6.2
flake8==3.6.0
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
parso==0.3.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.0
prompt-toolkit==2.0.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.10.0
pytz==2018.5
requests==2.20.0
six==1.11.0
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
wrapt==1.10.11
<<<<<<< HEAD
=======
whitenoise==4.1
>>>>>>> c0ac52dcb053bc85fb8dc2e92cd627173223d9cb
